Not allow to delete products if they are in another dessert

+++ pages/api/products/update.ts
@@ -1,7 +1,9 @@
 import type { NextApiRequest, NextApiResponse } from 'next';
+import mongoose from "mongoose";
 
 import connectDB from '../../../api/database/connectDB';
 import Product from '../../../api/modules/products/product.model';
+import Dessert from '../../../api/modules/desserts/dessert.model';
 
 const handler = async (req: NextApiRequest, res: NextApiResponse) => {
   const {
@@ -41,6 +43,12 @@ const handler = async (req: NextApiRequest, res: NextApiResponse) => {
 
     case 'DELETE':
       try {
+        const matchedDessert = await Dessert.findOne({ products: { $elemMatch: { product: new mongoose.Types.ObjectId(id as string) } } });
+
+        if (matchedDessert) {
+          return res.status(400).json({ success: false, message: `Продукт використовується в десерті: ${matchedDessert.name}` })
+        }
+
         const deletedProduct = await Product.deleteOne({ _id: id });
         if (!deletedProduct) {
           return res.status(400).json({ success: false });
